Add menu

+++ src/app/assignments/assignment-detail/assignment-detail.component.ts
@@ -5,51 +5,56 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { AuthService } from 'src/app/shared/auth.service';
 
 @Component({
-  selector: 'app-assignment-detail',
-  templateUrl: './assignment-detail.component.html',
-  styleUrls: ['./assignment-detail.component.css']
+    selector: 'app-assignment-detail',
+    templateUrl: './assignment-detail.component.html',
+    styleUrls: ['./assignment-detail.component.css']
 })
 export class AssignmentDetailComponent implements OnInit {
 
-  assignementTransmis: Assignment;
+    assignementTransmis: Assignment;
 
-  constructor(private assignementService: AssignmentsService, private route: ActivatedRoute, private router: Router, private authService: AuthService) { }
+    constructor(private assignementService: AssignmentsService, private route: ActivatedRoute, private router: Router, private authService: AuthService) { }
 
-  ngOnInit(): void {
-    this.getAssignement()
-  }
+    ngOnInit(): void {
+        this.getAssignement()
+    }
 
-  onDelete() {
-    this.assignementService.deleteAssignment(this.assignementTransmis).subscribe(message => {
-      console.log(message);
-      this.router.navigate(['home']);
-    })
-    
-  }
+    onClickDelete() {
+        let isLogged;
+        this.authService.isLoggedIn.subscribe(success => {
+            isLogged = success;
+        });
 
-  onAssignementRendu() {
-    this.assignementTransmis.rendu = true;
-    this.assignementService.updateAssignment(this.assignementTransmis).subscribe(message => {
-      console.log(message);
-      this.router.navigate(['home']);
-    });
-    
-  }
+        if (!isLogged) {
+            this.router.navigate(['login']);
+        } else {
+            this.assignementService.deleteAssignment(this.assignementTransmis).subscribe(message => {
+                console.log(message);
+                this.router.navigate(['home']);
+            })
+        }
 
-  getAssignement() {
-    const id = +this.route.snapshot.params['id'];
-    console.log('id : ', id);
-    this.assignementService.getAssignment(id).subscribe(assignement => {
-      this.assignementTransmis = assignement
-      console.log("Res ass detail ", assignement);
-    })
-  }
-  onClickEdit() {
-    this.router.navigate(['/assignement', this.assignementTransmis.id, 'edit'],
-    {queryParams:{nom:this.assignementTransmis.nom}, fragment:'edition'});
-  }
+    }
 
-  isAdmin() {
-    return this.authService.isAdmin();
-  }
+    onAssignementRendu() {
+        this.assignementTransmis.rendu = true;
+        this.assignementService.updateAssignment(this.assignementTransmis).subscribe(message => {
+            console.log(message);
+            this.router.navigate(['home']);
+        });
+
+    }
+
+    getAssignement() {
+        const id = +this.route.snapshot.params['id'];
+        console.log('id : ', id);
+        this.assignementService.getAssignment(id).subscribe(assignement => {
+            this.assignementTransmis = assignement
+            console.log("Res ass detail ", assignement);
+        })
+    }
+    onClickEdit() {
+        this.router.navigate(['/assignement', this.assignementTransmis.id, 'edit'],
+            { queryParams: { nom: this.assignementTransmis.nom }, fragment: 'edition' });
+    }
 }
